+++ CHANGELOG.md
@@ -6,7 +6,21 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Added
 
-- Nothing.
+- [#7](https://github.com/zendframework/zend-expressive-platesrenderer/pull/7)
+  adds:
+  - `Zend\Expressive\Plates\PlatesEngineFactory`, which will create and return a
+    `League\Plates\Engine` instance. It introspects the `plates.extensions`
+    configuration to optionally load extensions into the engine; that value must
+    be an array of:
+    - extension instances
+    - string service names resolving to extension instances
+    - string class names resolving to extension instances
+  - `Zend\Expressive\Plates\Extension\UrlExtension`, which provides a wrapper
+    around the `UrlHelper` and `ServerUrlHelper` from zend-expressive-helpers,
+    as the functions `url($route = null, array $params = []) : string` and
+    `serverurl($path = null) : string`, respectively.
+  - `Zend\Expressive\Plates\Extension\UrlExtensionFactory`, which provides a
+    factory for creating the `UrlExtension`.
 
 ### Deprecated
 
@@ -18,7 +32,11 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Fixed
 
-- Nothing.
+- [#7](https://github.com/zendframework/zend-expressive-platesrenderer/pull/7)
+  updates `PlatesRendererFactory` to use either the `League\Plates\Engine`
+  service, if available, or the new `PlatesEngineFactory` to create the Plates
+  engine instance. This also ensures the `url()` and `serverurl()` functions are
+  registered by default.
 
 ## 1.0.1 - TBD